refactor(model_estimation): transfer `fit.filter.py`s into _model_estimation.py_
